pyproject.toml
[project]
name = "tojibot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright",
    "gspread",
    "google-auth",
    "google-api-python-client",
    "python-dotenv",
    "requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tojibot"]
